+++ nvml-wrapper-sys/build.rs
@@ -12,5 +12,18 @@ fn main() {
 
 #[cfg(target_os = "linux")]
 fn main() {
+    let paths = std::fs::read_dir("/usr/lib").unwrap();
+
+    for path in paths {
+        let entry = path.unwrap().path();
+        if entry.is_dir() {
+            let entry_string = entry.to_string_lossy();
+            if entry_string.contains("nvidia") && !entry_string.ends_with("prime") {
+                println!("cargo:rustc-link-search=native={}", entry_string);
+                break;
+            }
+        }
+    }
+
     println!("cargo:rustc-link-lib=nvidia-ml");
 }
